requirements.txt
chess
joblib
numpy
pillow
pygame
pyperclip
pyvips
scikit-learn
scipy
tensorflow
